+++ src/Api/Api.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Copyright 2019 Cloud Creativity Limited
  *
@@ -25,17 +24,14 @@ use CloudCreativity\LaravelJsonApi\Codec\EncodingList;
 use CloudCreativity\LaravelJsonApi\Contracts\Client\ClientInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\ContainerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Encoder\SerializerInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Repositories\ErrorRepositoryInterface;
+use CloudCreativity\LaravelJsonApi\Contracts\Exceptions\ExceptionParserInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Resolver\ResolverInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Store\StoreInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Validators\ValidatorFactoryInterface;
 use CloudCreativity\LaravelJsonApi\Factories\Factory;
 use CloudCreativity\LaravelJsonApi\Http\Responses\Responses;
 use CloudCreativity\LaravelJsonApi\Resolver\AggregateResolver;
 use CloudCreativity\LaravelJsonApi\Resolver\NamespaceResolver;
 use GuzzleHttp\Client;
-use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
-use Neomerx\JsonApi\Contracts\Encoder\EncoderInterface;
 use Neomerx\JsonApi\Contracts\Http\Headers\MediaTypeInterface;
 use Neomerx\JsonApi\Contracts\Http\Headers\SupportedExtensionsInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
@@ -74,12 +70,6 @@ class Api
     private $decodings;
 
     /**
-     * @var array
-     * @deprecated 2.0.0
-     */
-    private $errors;
-
-    /**
      * @var bool
      */
     private $useEloquent;
@@ -108,17 +98,6 @@ class Api
      * @var StoreInterface|null
      */
     private $store;
-
-    /**
-     * @var CodecMatcherInterface|null
-     */
-    private $codecMatcher;
-
-    /**
-     * @var ErrorRepositoryInterface|null
-     * @deprecated 2.0.0
-     */
-    private $errorRepository;
 
     /**
      * @var Responses|null
@@ -152,7 +131,6 @@ class Api
      * @param Jobs $jobs
      * @param bool $useEloquent
      * @param string|null $supportedExt
-     * @param array $errors
      * @param array $providers
      * @param string|null $connection
      * @param bool $transactions
@@ -167,7 +145,6 @@ class Api
         Jobs $jobs,
         $useEloquent = true,
         $supportedExt = null,
-        array $errors = [],
         array $providers = [],
         string $connection = null,
         bool $transactions = true
@@ -181,7 +158,6 @@ class Api
         $this->jobs = $jobs;
         $this->useEloquent = $useEloquent;
         $this->supportedExt = $supportedExt;
-        $this->errors = $errors;
         $this->providers = $providers;
         $this->connection = $connection;
         $this->transactions = $transactions;
@@ -192,10 +168,8 @@ class Api
      */
     public function __clone()
     {
-        $this->schemas = null;
+        $this->container = null;
         $this->store = null;
-        $this->codecMatcher = null;
-        $this->errorRepository = null;
     }
 
     /**
@@ -285,19 +259,6 @@ class Api
     }
 
     /**
-     * @return ErrorRepositoryInterface
-     * @deprecated 2.0.0
-     */
-    public function getErrors()
-    {
-        if (!$this->errorRepository) {
-            $this->errorRepository = $this->factory->createErrorRepository($this->errors);
-        }
-
-        return $this->errorRepository;
-    }
-
-    /**
      * @return SupportedExtensionsInterface|null
      */
     public function getSupportedExtensions()
@@ -371,6 +332,15 @@ class Api
     public function hasTransactions(): bool
     {
         return $this->transactions;
+    }
+
+    /**
+     * @return ExceptionParserInterface
+     * @todo add this to config.
+     */
+    public function exceptions(): ExceptionParserInterface
+    {
+        return app(ExceptionParserInterface::class);
     }
 
     /**
@@ -465,7 +435,6 @@ class Api
     public function register(AbstractProvider $provider)
     {
         $this->resolver->attach($provider->getResolver());
-        $this->errors = array_replace($provider->getErrors(), $this->errors);
     }
 
 }