+++ src/Api/Repository.php
@@ -87,7 +87,6 @@ class Repository
             Jobs::fromArray($config['jobs'] ?? []),
             $config['use-eloquent'],
             $config['supported-ext'],
-            $config['errors'],
             $config['providers'] ?? [],
             $config['controllers']['connection'] ?? null,
             $config['controllers']['transactions'] ?? true
@@ -134,7 +133,6 @@ class Repository
 
         $config['resources'] = $this->normalizeResources($config['resources'] ?? [], $config);
         $config['url'] = $this->normalizeUrl($config['url'] ?? [], $host);
-        $config['errors'] = array_replace($this->defaultErrors(), $config['errors'] ?? []);
         $config['controllers'] = $config['controllers'] ?? [];
 
         return $config;
@@ -150,14 +148,6 @@ class Repository
         $key = "json-api-$apiName";
 
         return $path ? "$key.$path" : $key;
-    }
-
-    /**
-     * @return array
-     */
-    private function defaultErrors()
-    {
-        return (array) $this->config->get('json-api-errors');
     }
 
     /**