+++ src/Validation/Validator.php
@@ -18,6 +18,7 @@
 namespace CloudCreativity\LaravelJsonApi\Validation;
 
 use CloudCreativity\LaravelJsonApi\Contracts\Validation\ValidatorInterface;
+use CloudCreativity\LaravelJsonApi\Document\Error\Translator as ErrorTranslator;
 use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use Illuminate\Contracts\Validation\Validator as ValidatorContract;
 use Neomerx\JsonApi\Exceptions\ErrorCollection;